requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pyyaml>=6.0

# Data handling
jsonlines>=3.1.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Optional but recommended
accelerate>=0.20.0  # For efficient model loading
